+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -96,6 +96,11 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`mlir-op-builder
+  <clang-tidy/checks/mlir/op-builder>` check.
+
+  Flags usage of old OpBuilder format.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^